+++ esperanto-tools-libs/tools/src/Worker.h
@@ -17,8 +17,8 @@
 class Worker {
 public:
   explicit Worker(size_t bytesH2D, size_t bytesD2H, size_t numH2D, size_t numD2H, rt::DeviceId device,
-                  rt::IRuntime& runtime);
-  void start(int numIterations);
+                  rt::IRuntime& runtime, const std::string& kernelPath);
+  void start(int numIterations, bool computeOpStats);
   rt::IBenchmarker::WorkerResult wait();
   ~Worker();
 
@@ -30,10 +30,18 @@ private:
   rt::IBenchmarker::WorkerResult result_;
   size_t numH2D_;
   size_t numD2H_;
+  std::optional<rt::KernelId> kernel_;
 
   std::vector<std::byte> hH2D_;
   std::byte* dH2D_ = nullptr;
 
   std::vector<std::byte> hD2H_;
   std::byte* dD2H_ = nullptr;
+
+  struct Parameters {
+    std::byte* src;
+    size_t srcSize;
+    std::byte* dst;
+    size_t dstSize;
+  } parameters_;
 };