+++ esperanto-tools-libs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unrealeased]
 ### Added
 - [SW-15123] Added Start and Stop events to runtime traces. Added parentId field for commands derived from a single application-level eventId
+- [SW-14909] Added API method to set profiler in Server (runtime mp daemon)
 ### Changed
 - (Conan) Refactor sysemu artifacts provided when runtime:with_tests=True
 - (Conan) Update accepted range of deviceLayer versions from `>=1.1.0 <1.2.0` to `>=1.1.0 <2.0.0`